+++ sdk/highlight-run/src/sdk/record.ts
@@ -73,10 +73,8 @@ import {
 	setItem,
 } from '../client/utils/storage'
 import { getDefaultDataURLOptions } from '../client/utils/utils'
-import {
-	payloadToBase64,
-	type HighlightClientRequestWorker,
-} from '../client/workers/highlight-client-worker'
+import { type HighlightClientRequestWorker } from '../client/workers/highlight-client-worker'
+import { payloadToBase64 } from '../client/utils/payload'
 import HighlightClientWorker from '../client/workers/highlight-client-worker?worker&inline'
 import { MessageType, PropertyType } from '../client/workers/types'
 import { IntegrationClient } from '../integrations'
